+++ packages/terra-list/tests/jest/List.test.jsx
@@ -110,6 +110,51 @@ it('should render with ariaDetails', () => {
   expect(shallowComponent).toMatchSnapshot();
 });
 
+it('should render with ariaDescribedBy', () => {
+  const item1 = <Item key="123" />;
+  const item2 = <Item key="124" />;
+  const item3 = <Item key="125" />;
+  const item4 = <Item key="126" />;
+  const item5 = <Item key="127" />;
+  const items = [item1, item2, item3, item4, item5];
+  const shallowComponent = shallowWithIntl(
+    <div>
+      <p id="list-help"> Navigate this list using the arrow keys. </p>
+      <List ariaDescribedBy="list-help">{items}</List>
+    </div>,
+  );
+  expect(shallowComponent).toMatchSnapshot();
+});
+
+it('should render with ariaDescription', () => {
+  const item1 = <Item key="123" />;
+  const item2 = <Item key="124" />;
+  const item3 = <Item key="125" />;
+  const item4 = <Item key="126" />;
+  const item5 = <Item key="127" />;
+  const items = [item1, item2, item3, item4, item5];
+  const shallowComponent = shallowWithIntl(
+    <List ariaDescription="Navigate this list using the arrow keys.">{items}</List>,
+  );
+  expect(shallowComponent).toMatchSnapshot();
+});
+
+it('should render with ariaDetails', () => {
+  const item1 = <Item key="123" />;
+  const item2 = <Item key="124" />;
+  const item3 = <Item key="125" />;
+  const item4 = <Item key="126" />;
+  const item5 = <Item key="127" />;
+  const items = [item1, item2, item3, item4, item5];
+  const shallowComponent = shallowWithIntl(
+    <div>
+      <p id="list-details"> Here is some more information about this list. </p>
+      <List ariaDetails="list-details">{items}</List>
+    </div>,
+  );
+  expect(shallowComponent).toMatchSnapshot();
+});
+
 it('correctly applies the theme context className', () => {
   const wrapper = mountWithIntl(
     <ThemeContextProvider theme={{ className: 'orion-fusion-theme' }}>