+++ packages/terra-core-docs/CHANGELOG.md
@@ -19,10 +19,13 @@
 * Added
   * Added Higher Order examples for `terra-profile-image`.
   * Added Accessibility Guide for `terra-Profile-image`.
+
+* Changed
   * Added ariaLabel and ariaLabelledBy examples for `terra-toolbar`.
   * Added Accordion Examples for `terra-section-header`.
   * Added accessibility guide for `terra-divider`.
   * Added tests and example for adding row header to `terra-html-table`.
+  * Added `terra-list` tests and examples for `ariaDescribedBy`, `ariaDescription`, and `ariaDetails` props.
   * Added accessibility guide for `terra-html-table`.
   * Added an accessibility hooks example for `terra-form-field` and `terra-input`.